requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pdfplumber>=0.10.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
